refactor: HTTP 메시지를 감싸는 패키지명 수정
