+++ src/main/java/webserver/RequestDispatcher.java
@@ -1,9 +1,10 @@
 package webserver;
 
 import controller.UserController;
+import utils.DataConverter;
 import utils.FileLoader;
-import webserver.domain.request.Request;
-import webserver.domain.response.Response;
+import webserver.message.request.Request;
+import webserver.message.response.Response;
 
 import java.util.HashMap;
 import java.util.Map;
