+++ src/main/java/webserver/StaticFile.java
@@ -1,4 +1,4 @@
-package webserver.domain.response;
+package webserver;
 
 import utils.FileIoUtils;
