+++ src/main/java/webserver/message/response/ResponseBody.java
@@ -1,4 +1,4 @@
-package webserver.domain.response;
+package webserver.message.response;
 
 public class ResponseBody {
     private byte[] body = new byte[0];
